Fix for issue 15
if writeObject is used for writing then readObject has to be used for reading. If
writeUnshared() is used for writing, then readUnshared has to be used for reading.

Hazelcast serializer is updated to use in.writeUnshared() instead of
in.writeObject(). now writeUnshared is used for writing and readUnshared is used for
reading.

git-svn-id: http://hazelcast.googlecode.com/svn/trunk@133 3f8e66b6-ca9d-11dd-a2b5-e5f827957e07