Fix for bind event bug. It was eating CPU because of a recursive call. CONNECT event should triger sending the Bind message, not ACCEPT.
Fixes issue 111
Removes are delayed for 10 seconds so that record owner can keep the right version for a little longer to avoid version confict on the backup side.
This is also fixing the issue 111. here is why:
Upon migration, both record owner and record backup are sending migrate (own) request to the actual owner. if the record is removed immediately you can end up having
own(from old owner)+remove+own(from old backup)+remove 
because second 'own' will not know that the record is actually migrated (owned) before and it was also removed. 




git-svn-id: http://hazelcast.googlecode.com/svn/trunk@483 3f8e66b6-ca9d-11dd-a2b5-e5f827957e07