Improves AWS Logging.
A user is now able to see why an instance is not selected; either because of the state, tag not matching or groupname not matching.

Also the instance name is displayed next to the private ip address to make it easier to see which instance is or is not selected.