MAX_NO_MASTER_CONFIRMATION_SECONDS increased to 500s.
It decrease a chance of data loss when multiple members are running on the same host which dies suddenly.

There are two commonly used mechanisms to detect dead members:
 - heartbeats
 - master confirmation

By increasing the MAX_NO_MASTER_CONFIRMATION_SECONDS we make sure the dead node is always detected by
the heartbeat mechanism. Heartbeats are sent with higher frequency than the master confirmations so there
is a good chance we detect all dead members before re-partitioning process is started.