Remove false-negative assertion in partition lost listener stress test
Removed assertion is not reliable because the failure of the assertion does not indicate invalid behavior.
Consider the following scenario:
Lets say that a partition has the replica table of 6 Nodes as [N1, N2, N3, N4, N5, N6]. N1, N2, N3, N4, N5 fails and N6 keeps running.
- If N1 + N2 fails while N3 is still alive, N3 reports that the partition is lost with 1 backup and it becomes owner.
   After that, N3 + N4 + N5 fails and N6 reports that the partition is lost with 2 backups since it has become the 3rd replica after N1 + N2 failure.
- If N1 + N2 + N3 fails while N4 is still alive, N4 reports that the partition is lost with 2 backups and it becomes owner.
   After that, N4 + N5 fails and N6 reports that the partition is lost with 1 backup.
For both of these scenarios, all of the partition lost reports are valid but the removed assertion fails on the second scenario.

 * Contains other refactoring on test classes such as re-formatting logs and moving methods

 Fixes #5047