Balancer improvements:
1: not all connections were seen by the balancer. Some active connections
were not seen, but receive load. Because of these connections being invisble
the balancer doesn't deal with balancing load correctly.

2: if there is 1 selector with a huge number of events, but only a single
handler, this was accepted as a good source selector to move a handler from.
In case of a single handler it doesn't make any sense since you don't want
to move a single handler from an otherwise empty selector.

1: has been fixed by registering every active connection at the iobalancer instead
of only the 'registered' ones

2: has been fixed by ignoring selector with a single handler; even though it has a huge number
of events