Spurious test failure fixed. Fixing #5862
The test creates creates a cluster of 3 members, introduces a split-brain and then attempts to heal it.

Now the healing process has two-steps: first it calls prepareMerge() which schedules the actual merge. However the test also calls the same merge() method on its own. So at the end of the day there are 2 threads attempting to call the merge() method. The merge() is protected by a AtomicBoolean to ensure only 1 thread is doing a merge. In most cases it's the test thread which win the CAS. But if the scheduled thread wins then the test-thread will lose the CAS on and will return before the merge is completed -> Test Failure.