Fixes problem in client side logic ringbuffer
The problem was that the incorrect field was being used to determine the size of the result result set.
It should use the method size(), but it relied on an internal array which can have a different size (like
array in arraylist).

Also the ClientReliableTopicTest is now made a normal test instead of a nightly.