Provides fast `Set` implementation for cases where items are known to not contain duplicates.
It doesn't not call equals/hash methods on initial data insertion hence it avoids
performance penalty in the case these methods are expensive. It also means it does
not detect duplicates - it's a responsibility of a caller to make sure no duplicated
entries are inserted.

Once the initial load is done then caller should indicate it
by calling `close()`. It switches operational mode and if any item is
inserted afterward then the Set will inflate - it copies its content into
internal HashSet. This means duplicate detection is enabled, but obviously
it ruins the initial performance gain. We are making a bet the Set will not be
modified once it's closed. The Set will also inflate on `contains()`
and `containsAll()` calls to enable fast look-ups.