Renamed ReadHandler and WriteHandler
to SocketReader and SocketWriter. The original names are flipped; the socket
reader process the read data, and the read handler reads data from the socket.

The same goes for the writing part.

So in this PR these names are flipped so that the SocketReader is responsible
for reading from the socket and the ReadHandler is process the incoming data.