NotEquals predicate is again just a negated Equals predicate.
In other words: not(equals(p)) <-> notEquals(p)
Fixes #6124

It's partially reverting a changeset #5023.
The original change was introduced to treat NULL-values more like relational databases do.
SQL standard mandates special treating of NULL values, but it order to be compliant we'd
need to introduce three-values logic as specified in SQL. Having just a single predicate
behaving like SQL mandates creates inconsistencies and logical paradoxes.

The #5023 was also fixing a NPE when NotEquals used an index, however this revert will not
re-introduce the NPE  indexes are no longer used for NotEquals predicate. See #5847