Introduce ClusterState
Cluster will have 3 states; ACTIVE, FROZEN and SHUTTIND_DOWN.

- ACTIVE is normal operating state of the cluster. All operations will be allowed to run,
 nothing will be restricted in the cluster.
- FROZEN is to support rolling upgrade. In FROZEN state, cluster members
and partition assignments will be frozen. No new member will be allowed to join,
only the ones left in this state will be allowed.
- SHUTTING_DOWN is shutdown preparation state. This state add more restrictions to the FROZEN state.
Additional to the FROZEN state, no operations will be allowed to run in the cluster,
except the ones marked as AllowedDuringShutdown.