Make sure the cluster uptime > 0 when running PhoneHomeTest
Forward-port of #7197, Fixes #7192

Reasoning:
The test assumes the cluster-uptime is always greater than 0 (ms).
The cluster uptime is defined as "no. of ms since cluster has started".
However if the test is too fast then the cluster time is 0 -> the test
is failing.

I am also reverting b170651b3 as it has no effect here and it's
effectively noise only.