Invalid partition id should not crash partition threads
When an operation, a partition-specific-runnable or a packet with an
invalid partition id (greater than max partition id) is submitted to
operation executor, it causes partition operation thread to crash with
ArrayIndexOutOfBoundsException. After that point, no task can be executed
on any that thread and also thread's task queue can lead to oome.

Even this can be considered a security failure, one can send a corrupted
client message to a server and server becomes stuck.