Cleaned up HazelcastProperties
Removed the need to create an array by letting the subclass implement a method.
Everything is now done through the constructor.

The HazelcastProperties now also have abilit to ask the keyset + value. This is
useful to determine the content of the HazelcastProperties.

Fixed the ConfigPropertiesPlugin to make use of these new methods.