Properties are now more flexible.
Properties are currently inflexible + lot of duplication. Now properties don't
need to be defined on the either Client or GroupProperty; but they can just
be used. This allows for a more flexible approach.

One of the reason to clean this up is that the metrics were using group property,
and this caused big problems when using them on a client.