Added probes to let handlers show their thread.id
Each handler is run on an io thread and with these probes in place,
in the metrics one can see which iothread a handler is using. This
is very useful to figure out of there are any imbalances.

Also a migration counter has been added so we can see how many times
a socketreader/writes has been migrated