Solution a race when a client connection is closed
When connection closed, it was possible to endup with invocations
that are not notified. These invocations was causing infinite
wait. We let the race happen and check the callIdMap from another
thread every seconds to cleanup unlucky invocations.

There was a workaround for this in ClientInvcationFuture.get
which is waking up every hearbeatTimeout milliseconds and checking
if connection is still there. This workaround is removed in a
previous pr because when async api is used that workaround is
does not work.

fixes #7582 , #7527.