fix to a regression caused by #7635.
Invocations of dead connections were handling as soon as connection
is closed before #7635. With the change in #7635, the invocations
are started to handled by a separate thread periodically.
Since that thread is closed when client itself closes, there can
be some invocations left unhandled. This pr makes sure they are
handled in ClienInvocationService shutdown.

fixes #7527