Complete new migration implementation, final touches
- Add address->uuid mapping table to partition state.
When a member is restarted in a short period, before it's removed
from partition table in time, we need to identify that an old member is
removed and started back as a new member. This is required to be able
to repair partition table correctly.
A mapping for address->uuid is kept alongside with partition
table and during repair we lookup a member's uuid in that mapping
table. If it's not found, it means that member is a new member.

- Clean up FinalizeMigrationOperation
- Extract reference impl of MigrationAwareService
- Split MigrationCorrectnessTest into multiple tests
- Partition package improvements
	* Added metrics registrations
	* Added javadocs
- Fix isReplicaInSyncState() checks
	* If there is at least one data node, all partitions must have an owner
- RepartitioningTask should pause when cluster is not active
- Count currently running migration system task while checking migration queue
- Remove partition state version increment when member added or removed
- After partitioning & migration process changes, version increment on
membership change is not needed.
- Remove slow correctness tests, add their params to the quick ones
- Fix member information sent by master alongside partition-table
	* Master should send its address-uuid mapping with partition-table,
since this mapping is a part of partition-table.
	* Also, master should update address-uuid mapping after a member joins
when cluster state is FROZEN or PASSIVE.