InvocationContext introduced
The goal of the InvocationContext is make the invocaiton a bit less dependant on the OperationService. Currently
the OperationService is used as 'context' to get dependencies from and if you want an Invocation, you need an
OperationServiceImpl -> NodeEngineImpl -> Node. So you can't use Invocations without having a cluster up and running.

The InvocationContext is the object with all dependencies for the invocations. Apart from reducing the need for an
OperationService, a lot of trainwrecks have been removed by adding the dependency directly on the InvocationContext.
This used to be a problem because not all dependencies exist when the OperationService is constructed, but the
InvocationContext is created after these dependencies are created, so the InvocationContext can get these
dependencies.