OperationTimeoutException improvement
Includes the current time in the exception. On multiple occassions I have seen this stacktrace and could
not figure out the current time easily since only this exception is being copied into error message.

So it now includes the Clock time. This makes interpreting the other 'time' values a lot easier.