OperationServiceImpl.thisAddress added
To reduce coupling to nodeengine, a thisAddress field is added to OperationService.
So during construction there is a dependency on Node for thisAddress, but once
constructed, the node is not needed to provide the thisAddress. This makes it
easier in the future get get rid of reliance on node/nodeengineiml