Use return value from TestHazelcastInstanceFactory.initOrCreateConfig
When `null` config is passed to `TestHazelcastInstanceFactory.initOrCreateConfig()`,
it creates a new one returns it, otherwise initializes the config. But if returned
config is not used, a new default config is created while starting a new Hazelcast
instance and that causes failures on build system because multicast is enabled
in default config and that breaks isolation between tests.