Remove partition backup version assertion in graceful shutdown test
During a shift down migration, backup operations are missed between the destination's commit and the source's commit, which is not an issue of the migration system. It is about how our replication mechanism works so basically it is related to the write consistency work.

This issue is normally fixed by the anti-entropy system. Since anti-entropy is disabled within migration tests, we do not check backup states. We only check if the data resides in the partition owner.