fix #945 xml multicast disabled by default
This is a inconsistency with the xsd; which clearly states that xml is enabled by default.
It also is an inconsistency with the programmatic api, where multicast is enabled by default.

This fix resolves these inconsistencies. However it could be that people that only configure
tcp and have not explicitly disabled multicast, since it was disabled by default, now get an
error because multicast needs to be explicitly disabled (you can have multiple join mechanisms active).