Fixes sync_whenOperationTimeout
The cause of the test failure is a fire low callTimeout. Since the tests are
run in parallel and there is little control on the environment, one can easily
run into pauses of the jvm that lead to this test to fail.

So increased the timeout to make the test less sensitive to such failures.