Fix missing MigrationAwareService.beforeMigration() calls
`MigrationAwareService.beforeMigration()` should be called for each service,
on both source and destination sides and there must be an exactly matching `commitMigration()`
or `rollbackMigration` for each `beforeMigration()` call.

But under some cases, there were missing `beforeMigration()` calls, which breaks this guarantee.
These cases are;
- when another `MigrationAwareService`'s `beforeMigration()` throws an exception
- when a `MigrationAwareService` doesn't provide replication-operation to be
executed on destination side

Now, these missing calls are fixed and current tests are enhanced to check this requirement
for `beforeMigration()`, `commitMigration` and `rollbackMigration()` calls.