Fix REST shutdownNode test
When node is requested to be shutdown via REST command, response may not be
sent to the caller, if connection is closed before writing response
to the socket channel. In that case, caller will get a connection exception
instead of actual response.

Also, changed shutdown command to use `HazelcastInstance.shutdown()` instead
of `Node.shutdown()`.

Fixes #8106