Delay operation retry when node is not joined Fixes #3754, #3395
Reasoning:
When a smaller cluster is being merged into a bigger one then
operations waiting in the WaitNotify queues are cleared and
pending invocations receive MemberLeftException. However the
invocation can retry immediately and send the operation again.

This change is adding a check into a retry part to re-execute the
operation only when a node is joined (again). Happy path is
unaffected.

This commit is also adding a support class for split-brain tests.
Many thanks to @mdogan who helped me with analysis and also contributed
the original test.