Priority client messages executed on generic operation executor
Normally all generic client messages are processed by the client executor. However there is no priority,
and tasks in the executor can get processed very slowely. This can cause stability problems e.g. by
a failure to respond to heartbeats in time.

This fix offloads the most important priority generic messages to the operatione executor where there is a
dedicated thread dealing with priority tasks. This makes it a lot less unlikely that the system can't reply
to a heartbeat in time.