SlowOperation should block a partition thread just enough to cause Call-Timeout
Fixes #7932
Reasoning: We want the `SlowOperation` to block just slightly over the call-timeout.
This is sufficient to get the `DummyOperation` refused with call-timeout exception.

Having `SlowOperation` too slow means the invocation with `DummyOperation` will
fail due lack of heartbeat.

In other words: We want Invocation System to actively refuse `DummyOperation` to
be executed.