Fixed wrong calculation of Near Cache owned memory costs
All Near Caches implementations which base on the AbstractNearCacheRecordStore suffered from a wrong calculation of owned memory costs. Whenever a record in the Near Cache was updated, we upfront added the memory costs for its key + value. If an old records existed, we detracted just the costs for the value, but not the key. If the record was removed, we detracted the costs for key + value again. So in the end each record update piled up the costs of a key.