PartitionIteratingOperation & Backup more detail in diagnostics
If a PartitionIteratingOperation or a Backup is found in the invocations or
in the pending connections, instead of just mentioning the outer classname,
the diagnostics tries to look inside so one can see e.g.
PartitionIteratingOperation(MapSizeFactory) as name.

This gives better insight of the type of operations being encountered.

fix #8855