Fixes char array growining issue in Diagnostics
If a plugin has a lot of data to write, there are 2 buffers that are going to grow and
keep that size ever after; a stringbuffer and char array in the diagnostics sytem. This
is unacceptable since it can lead to unwanted memory usage.

This fix addresses the issue by immediately writing to a PrintWriter instead of
first to a strinbuffer then to an array and then to file.