Client io system makes use of ReadHandler
This is one other step to allign the client with the server. The idea is that we have
one non blocking socket reader/writer which can be used for both client and server.

The main change in this pr is that the ClientMessageBuilder is now renamed to
ClientMessageReadHandler and implements the ReadHandler interface from the server.